feat: move file to dir
