+++ src/components/SubscribeItem/Member/index.ts
@@ -1,0 +1,3 @@
+import Member from './Member';
+
+export default Member;
