+++ src/components/SubscribeItem/Member/style.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/native';
+import { ThemeProps } from 'theme';
+
+export const MemberView = styled.View<ThemeProps>`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  width: 90px;
+  padding-right: 20px;
+`;
