+++ src/components/SubscribeItem/MemberPayStatus/index.ts
@@ -1,0 +1,3 @@
+import MemberPayStatus from './MemberPayStatus';
+
+export default MemberPayStatus;
