+++ src/components/SubscribeItem/PayStatus/index.ts
@@ -1,0 +1,3 @@
+import PayStatus from "./PayStatus";
+
+export default PayStatus
